Resolve S09P12C202-400 "/feat"

See merge request s09-webmobile1-sub2/S09P12C202!121

+++ backend/planty-service/src/main/java/com/planty/api/subscribe/response/UserSubscribeDatailResponse.java
@@ -13,7 +13,7 @@ import java.util.List;
 @ToString
 public class UserSubscribeDatailResponse {
     @NotNull
-    private Integer sid;
+    private Long sid;
 
     @NotNull
     private String startDate;
